Testing

+++ file_io/0-read_textfile.c
@@ -14,7 +14,7 @@
 ssize_t read_textfile(const char *filename, size_t letters)
 {
 	ssize_t o;
-	int w, r;
+	ssize_t w, r;
 	char *buffer;
 	
 	if (!filename)
